# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "DetectionOut": {
        "description": "One detection row, denormalized for the frontend Log type.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "log_source_id": {
            "title": "Log Source Id",
            "type": "string"
          },
          "log_source_name": {
            "title": "Log Source Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "relevance": {
            "default": 0,
            "title": "Relevance",
            "type": "integer"
          },
          "sample_fields": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Fields"
          },
          "techniques": {
            "items": {
              "$ref": "#/components/schemas/TechniqueRef"
            },
            "title": "Techniques",
            "type": "array"
          }
        },
        "required": [
          "id",
          "log_source_id",
          "log_source_name",
          "event_id",
          "name",
          "techniques"
        ],
        "title": "DetectionOut",
        "type": "object"
      },
      "ErrorOut": {
        "description": "Forma estándar de error de la API (ARCHITECTURE.md §3).\n\nNunca se devuelve 200 con un error dentro. `code` es estable y pensado para\nque el cliente ramifique sobre él; `detail` es para humanos.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "keys": {
            "items": {
              "type": "string"
            },
            "title": "Keys",
            "type": "array"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PaginatedDetections": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DetectionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "PaginatedDetections",
        "type": "object"
      },
...
      "TechniqueRef": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tactic": {
            "items": {
              "type": "string"
            },
            "title": "Tactic",
            "type": "array"
          },
          "technique_id": {
            "title": "Technique Id",
            "type": "string"
          },
          "technique_name": {
            "title": "Technique Name",
            "type": "string"
          }
        },
        "required": [
          "technique_id",
          "technique_name",
          "id",
          "name",
          "tactic",
          "confidence"
        ],
        "title": "TechniqueRef",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Log Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/logs": {
      "get": {
        "description": "`GET /logs?filter[<key>]=<value>&q=...&limit=...&cursor=...`\n\nLas claves de filtro no se enumeran aquí: se descubren del catálogo en\nruntime. Una clave desconocida devuelve 400 (`unknown_filter_category`);\nun cursor manipulado o de otro backend, 400 (`invalid_cursor`).",
        "operationId": "listLogs",
        "parameters": [
          {
            "description": "Free-text search",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Free-text search",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a previous next_cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous next_cursor",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDetections"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown filter category or bad cursor"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Logs",
        "tags": [
          "logs"
        ]
      }
    },
...
  }
}